requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pinecone-client
voyageai
python-dotenv
requests